+++ src/components/Dialog/dialog.less
@@ -2,6 +2,7 @@
 
 .overlay{
   display: none;
+  z-index: 1000;
   background-color: rgba(@black-color, .4);
   width: 100%;
   height: 100%;
