+++ src/pages/profile/changePassword/changePassword.hbs
@@ -1,6 +1,7 @@
 <main class="profile" id="changePassword-form">
   {{{ sidebar }}}
   <div class="profile__content">
+    {{{ profileHeader }}}
     <form class="profileChangePasswordForm">
       {{{ oldPasswordField }}}
       {{{ newPasswordField }}}
