+++ src/pages/profile/components/ProfileHeader/ProfileHeader.ts
@@ -1,24 +1,96 @@
 import Block from "/src/utils/Block";
 import template from "./profileHeader.hbs";
 import styles from "./profileHeader.less";
+import defaultAvatar from "/static/user_avatar.svg";
+import { TState } from "/src/types";
+import store from "/src/utils/Store";
+import { resourcesReducer, userReducer } from "/src/reducers";
+import ResourcesController from "/src/controllers/ResourcesController";
+import { Button, Dialog, InputFiled } from "/src/components";
+import { validateForm } from "/src/utils/validation/validateForm";
+import UserController from "/src/controllers/UserController";
+import { Avatar } from "/src/pages/profile/components/ProfileHeader/components/Avatar";
 
-interface IProfileHeader {
-  avatar: string;
-  first_name: string;
-  events: {
-    click: () => void;
+export class ProfileHeader extends Block {
+  private userData: TState;
+  private userAvatar = '';
+
+  protected initChildren() {
+    const state = store.getState();
+
+    this.userData = userReducer(state);
+
+    ResourcesController.getResources(this.userData.avatar);
+
+    this.userAvatar = resourcesReducer(state);
+
+    this.children.changeAvatarDialog = new Dialog({
+      title: 'Загрузите файл',
+      content: new InputFiled({
+        type: 'file',
+        id: 'avatar',
+        label: '',
+        name: 'avatar',
+        value: this.props?.avatar,
+        accept: 'image/*',
+        customClass: 'fileInput'
+      }),
+      submit: new Button({
+        label: 'Сохранить',
+        type: 'submit',
+        events: {
+          click: () => this.onChangeAvatarSubmit()
+        }
+      }),
+      cancel:  new Button({
+        label: 'Отменить',
+        type: 'button',
+        isGhost: true,
+        customClass: 'dialog__actions-cancel',
+        events: {
+          click: () => this.closeDialog()
+        }
+      }),
+    })
+
+    this.children.avatar = new Avatar({
+      avatarPath: this.userAvatar || defaultAvatar,
+      events: {
+        click: () => this.showDialog()
+      }
+    })
   }
-}
 
-export class ProfileHeader extends Block<IProfileHeader> {
-  constructor(props: IProfileHeader) {
-    super(props);
+  onChangeAvatarSubmit() {
+    const form = document.querySelector('#dialog-form') as HTMLFormElement;
+
+    form.onsubmit = (e) => {
+      e.preventDefault();
+
+      const data = new FormData(e.target as HTMLFormElement)
+
+      if (validateForm(data)) {
+        UserController.changeAvatar(data);
+
+        this.closeDialog();
+      }
+    }
+  }
+
+  showDialog() {
+    (this.children.changeAvatarDialog as Dialog).show();
+  }
+
+  closeDialog() {
+    (this.children.changeAvatarDialog as Dialog).hide();
   }
 
   protected render() {
     return this.compile(template, {
-      avatar: this.props.avatar,
-      first_name: this.props.first_name,
+      first_name: this.userData.first_name,
+      events: {
+        click: () => this.showDialog
+      },
       styles
     });
   }
