+++ src/pages/profile/components/ProfileHeader/components/avatar.hbs
@@ -1,0 +1,1 @@
+<img class="avatar__image" src={{ avatarPath }} alt="user__avatar">
