+++ src/pages/profile/components/ProfileHeader/components/avatar.less
@@ -1,0 +1,10 @@
+.avatar__image {
+  cursor: pointer;
+  border-radius: 50%;
+  width: 130px;
+  height: 130px;
+
+  &:hover {
+    opacity: 0.7;
+  }
+}
