+++ src/pages/profile/components/ProfileHeader/profileHeader.hbs
@@ -1,6 +1,7 @@
 <div class="profile__header">
   <div class="header__avatar">
-    <img class="avatar__image" src={{ avatar }} alt="user__avatar">
+    {{{ avatar }}}
   </div>
   <div class="header__name">{{{ first_name }}}</div>
+  {{{ changeAvatarDialog }}}
 </div>
