+++ src/pages/profile/components/ProfileHeader/profileHeader.less
@@ -2,17 +2,6 @@
   display: inline-block;
 }
 
-.avatar__image {
-  cursor: pointer;
-  border-radius: 50%;
-  width: 130px;
-  height: 130px;
-
-  &:hover {
-    opacity: 0.7;
-  }
-}
-
 .header__name {
   font-size: 16px;
   font-weight: 600;
