+++ src/pages/profile/edit/ProfileEditPage.ts
@@ -6,6 +6,7 @@ import styles from './profileEdit.less';
 import { TState } from '/src/types';
 import store from '/src/utils/Store';
 import { userReducer } from '/src/reducers';
+import { ProfileHeader } from "/src/pages/profile/components/ProfileHeader/ProfileHeader";
 
 export class ProfileEditPage extends Block {
   private userData: TState;
@@ -18,6 +19,8 @@ export class ProfileEditPage extends Block {
     this.children.sidebar = new ProfileSidebar();
 
     this.children.profileEditForm = new ProfileEditForm(this.userData);
+
+    this.children.profileHeader = new ProfileHeader();
   }
 
   protected render() {
