+++ src/pages/profile/edit/profileEdit.hbs
@@ -1,6 +1,7 @@
 <main class="profile">
   {{{ sidebar }}}
   <div class="profile__content">
+    {{{ profileHeader }}}
     {{{ profileEditForm }}}
   </div>
 </main>
